vocal_model/requirements.txt
torch
numpy
soundfile
librosa
tqdm
pyyaml
omegaconf
safetensors
einops
beartype
rotary_embedding_torch
